requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
py-cord @ git+https://github.com/Pycord-Development/pycord@0b54eee722f39c1961d83ea4019548e67654f366
python-dotenv==0.19.2
typing-extensions==4.1.1
yarl==1.7.2
